// src/app/Models/CountryInfo.Model.ts
export class CountryInfo {
    countryId: number = 0;
    countryName: string = "";
    iso2: string = "";
    iso3: string = "";
    countryCode: string = "";
    capitalCity: string = "";
    countryPopulation: string = "";
    continent: string = "";
}

export class CountryDropdown{
    countryId: number = 0;
    countryName: string = "";
}